+++ tests/tests_unit/test_api/test_assets.py
@@ -109,6 +109,14 @@ class TestAssets:
         assert "bla" == jsgz_load(mock_assets_response.calls[0].request.body)["filter"]["name"]
         assert mock_assets_response.calls[0].response.json()["items"] == res.dump(camel_case=True)
 
+    def test_list_ag(self, mock_assets_response):
+        res = ASSETS_API.list(name="bla",aggregated_properties=["childCount"])
+        assert ["childCount"] == jsgz_load(mock_assets_response.calls[0].request.body)["aggregatedProperties"]
+
+    def test_list_ag_camel(self, mock_assets_response):
+        res = ASSETS_API.list(name="bla",aggregated_properties=["child_count"])
+        assert ["childCount"] == jsgz_load(mock_assets_response.calls[0].request.body)["aggregatedProperties"]
+
     def test_create_single(self, mock_assets_response):
         res = ASSETS_API.create(Asset(external_id="1", name="blabla"))
         assert isinstance(res, Asset)